// src/profesional-profiles/enum/languages.enum.ts
export enum Languages {
  NODEJS = 'node.js',
  NODEJS_ALT = 'nodejs',
  JAVASCRIPT = 'javascript',
  TYPESCRIPT = 'typescript',
  CSHARP = 'c#',
  CPLUSPLUS = 'c++',
  C = 'c',
  PASCAL = 'pascal',
  FORTRAN = 'fortran',
  RUBY = 'ruby',
  PERL = 'perl',
  HTML = 'html',
  HTML_ALT = 'html5',
  PHP = 'php',
  JAVA = 'java',
  KOTLIN = 'kotlin',
  SQL = 'sql',
  CSS = 'css',
  CSS_ALT = 'css3',
  XML = 'xml',
  JSON = 'json',
}